requirements.txt
torch
transformers
python-chess
minicons
